feat: added/changed exception logic for services and repository

+++ KudaGo/src/main/java/arden/java/kudago/service/impl/LocationServiceImpl.java
@@ -1,6 +1,8 @@
 package arden.java.kudago.service.impl;
 
 import arden.java.kudago.dto.Location;
+import arden.java.kudago.exception.CreationObjectException;
+import arden.java.kudago.exception.IdNotFoundException;
 import arden.java.kudago.repository.StorageRepository;
 import arden.java.kudago.service.LocationService;
 import lombok.RequiredArgsConstructor;
@@ -22,21 +24,41 @@ public class LocationServiceImpl implements LocationService {
 
     @Override
     public Location getLocationBySlug(String slug) {
+        if (locationStorage.read(slug) == null) {
+            throw new IdNotFoundException("Location with slug '" + slug + "' not found");
+        }
+
         return locationStorage.read(slug);
     }
 
     @Override
-    public Location createLocation(Location Location) {
-        return locationStorage.create(Location.slug(), Location);
+    public Location createLocation(Location location) {
+        if (locationStorage.create(location.slug(), location) == null) {
+            throw new CreationObjectException("Could not create location, because your input format is wrong, check again");
+        }
+
+        return locationStorage.create(location.slug(), location);
     }
 
     @Override
-    public Location updateLocation(String slug, Location Location) {
-        return locationStorage.update(slug, Location);
+    public Location updateLocation(String slug, Location location) {
+        if (locationStorage.read(slug) == null) {
+            throw new IdNotFoundException("Location with slug '" + slug + "' not found");
+        }
+
+        if (locationStorage.update(slug, location) == null) {
+            throw new CreationObjectException("Could not update location, because your input format is wrong, check again");
+        }
+
+        return locationStorage.update(slug, location);
     }
 
     @Override
     public boolean deleteLocation(String slug) {
+        if (locationStorage.read(slug) == null || !locationStorage.delete(slug)) {
+            throw new IdNotFoundException("Location with slug '" + slug + "' not found");
+        }
+
         return locationStorage.delete(slug);
     }
 }
